converted React components from .js files into .jsx
